+++ src/lib.rs
@@ -20,9 +20,12 @@ use std::borrow::Cow;
 
 pub use client::Client;
 pub use error::Error;
+pub use events::Aggregate;
 pub use events::Event;
 pub use events::Events;
+pub use events::Quote;
 pub use events::Stock;
 pub use events::Subscription;
+pub use events::Trade;
 
 type Str = Cow<'static, str>;